+++ HammerModel/Model/Entities/Teleport.cs
@@ -11,12 +11,12 @@ namespace HammerModel.Model.Entities
 {
     public class Teleport : WorldObject
     {
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
-        public int Width { get; set; }
-        public int Breadth { get; set; }
-        public int Height { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
+        public double Width { get; set; }
+        public double Breadth { get; set; }
+        public double Height { get; set; }
         public string Target { get; set; }
 
         public override List<HammerObject> ToHammerObject()
